move sse to separate handler, add tests for configuration options, cleanup
